fetching data with server side, without rls
RLS (row level security) is for ROLE BASE ACCESS

with static folder, pls ignore. its my guide
- jo
