+++ package.json
@@ -21,6 +21,7 @@
     "@radix-ui/react-toast": "^1.2.14",
     "@radix-ui/react-tooltip": "^1.2.7",
     "@supabase/auth-helpers-nextjs": "^0.10.0",
+    "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.51.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
