+++ src/app/landing/clinical-instructor/page.tsx
@@ -51,7 +51,7 @@ export default function InstructorLoginPage() {
         console.log("Login success!", data)
         setErrorMessage("")
         setHasError(false)
-        router.push("/dashboard/clinician")
+        router.push("/dashboard/clinical-instructor")
       }
         
     }
